Added support for 0.1 as precision value. Changed data type from int to float.

+++ Resources/views/backend/config/model/form/currency.js
@@ -3,6 +3,6 @@
 //{$smarty.block.parent}
 { name: 'active', type: 'boolean' },
 { name: 'alwaysUp', type: 'boolean' },
-{ name: 'precision', type: 'int' },
+{ name: 'precision', type: 'float' },
 { name: 'subtrahend', type: 'float' },
 //{/block}
